requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.17

numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
sympy>=1.10.0

torchmetrics>=0.11.0

optuna>=3.0.0

tensorboard>=2.10.0
tqdm>=4.64.0
plotly>=5.11.0
matplotlib>=3.5.0
seaborn>=0.12.0
ipython>=8.0.0

pyyaml>=6.0

accelerate>=0.20.0

scikit-learn>=1.0.0
pymatgen>=2022.0.0
ase>=3.22.0
jarvis-tools>=2021.0.0
